fixed missing controls with try catch for auc

+++ slr_analyses/Data/sCD14_mse.R
@@ -86,17 +86,25 @@ res = foreach(
   #     -- since response is continuous, no need to stratify by case-control.
   numObs = nrow(X_gbm)
   inputDim = ncol(X_gbm)
-  trainIdx = sample(cut(1:numObs, breaks=5, labels=F))
-  XTr = X_gbm[trainIdx != 1,]
-  YTr = Y[trainIdx != 1]
-  XTe = X_gbm[trainIdx == 1,]
-  YTe = Y[trainIdx == 1]
-  
-  saveRDS(list(
-    XTr = XTr, YTr = YTr,
-    XTe = XTe, YTe = YTe
-  ),
-  paste0(output_dir, "/data", file.end))
+  if(file.exists(paste0(output_dir, "/data", file.end))){
+    data.tmp = readRDS(paste0(output_dir, "/data", file.end))
+    XTr = data.tmp$XTr
+    XTe = data.tmp$XTe
+    YTr = data.tmp$YTr
+    YTe = data.tmp$YTe
+  } else{
+    trainIdx = sample(cut(1:numObs, breaks=5, labels=F))
+    XTr = X_gbm[trainIdx != 1,]
+    YTr = Y[trainIdx != 1]
+    XTe = X_gbm[trainIdx == 1,]
+    YTe = Y[trainIdx == 1]
+    
+    saveRDS(list(
+      XTr = XTr, YTr = YTr,
+      XTe = XTe, YTe = YTe
+    ),
+    paste0(output_dir, "/data", file.end))
+  }
   
   ##############################################################################
   # fit methods
